health-clock-backend/requirements.txt
# FastAPI 核心
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# 数据验证
pydantic==2.9.2
pydantic-settings==2.6.0

# 数据库
supabase==2.9.0
asyncpg==0.30.0

# AI 框架
langchain==0.3.7
langchain-community==0.3.5
dashscope==1.20.9

# 对象存储
boto3==1.35.36

# 工具库
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.27.2

# 阿里云号码认证服务（Dypnsapi 2017-05-25）
alibabacloud_dypnsapi20170525==2.0.0
alibabacloud_tea_openapi==0.4.4
